requirements.txt
alembic==1.12.1
Flask==2.3.3
Flask-Admin==1.6.1
Flask-JWT-Extended==4.5.3
flask-login==0.6.3
Flask-Mailman==1.0.0
Flask-Migrate==4.0.5
flask-restful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
markupsafe==2.1.3
httpx==0.25.1
jsonschema==4.19.2
WTForms==3.1.1

# Code quality
black==23.10.1
flake8==6.1.0
pytest==7.4.3

# Build
setuptools==68.2.2
wheel==0.41.3
